afxdp/neighbor: resolve logical ifindex only on the learn path (#2370 review)
Review fold on PR #2425, three items.

1. Hot-path regression (primary). stage_link_layer_classify runs
   per-packet for ALL ingress traffic — it is invoked before the
   flow-cache fast path, so every frame passes through it. The #2370 fix
   computed `learn_ifindex = resolve_ingress_logical_ifindex(..)`
   unconditionally at the top of the stage, which made every
   non-ARP/non-NDP packet (the overwhelming majority, including
   flow-cache hits) pay a FastMap lookup that is only needed by the
   ARP-reply and NDP-NA learn branches. Pre-#2370 a plain data packet did
   zero resolve calls.

   Wrap the resolve in a `learn_ifindex` closure and call it ONLY inside
   the two learn arms. A non-ARP/non-NDP packet — classify_arp returns
   NotArp and parse_ndp_neighbor_advert returns None — never invokes the
   closure, so the FastMap lookup is skipped entirely and the pre-#2370
   zero-cost data path is restored. The learn-path result is unchanged
   (same logical ifindex keys both dynamic_neighbors.insert and
   add_kernel_neighbor); the VLAN fail-on-revert tests still prove the
   key is the logical one (3 fail on revert to the physical ifindex,
   untagged stays green).

2. Test drift. The ICMPv6 checksum stamper was duplicated between
   parser_tests.rs (used to stamp valid #2368 NDP NA frames) and the new
   poll_stages.rs #2370 NDP test. Factor it into one shared
   afxdp::test_fixtures::{compute_icmpv6_checksum, stamp_icmpv6_checksum}
   and use it from both sites so the pseudo-header / fold logic cannot
   drift.

3. Doc fix. The neighbor_learn_ctx Box::leak comment claimed "each test
   runs once and the process exits"; a single test binary runs many
   tests in one process, so the leaks accumulate. Corrected the comment
   to state the leak persists (bounded, test-only) for the test-binary
   lifetime.

cargo build --release clean; poll_stages 6/6 (incl. the NDP test through
the shared stamper), parser 31/31, combined
neighbor/dynamic/ifindex/vlan/poll_stages/forwarding 300 green; 5x
stable. No wire change.
